requirements.txt
#dotenv
python-dotenv

# web tools
flask[async]
flask
markdown
html-sanitizer

flask-socketio
requests

# Data import
pandas # Data manipulation library
Pillow # PIL image-loading library
beautifulsoup4 # Text scraper from websites 

# Math and data visualization
numpy
matplotlib
seaborn

#Data preprocessing and model builing for ML
nltk # Natural Language Toolkit
scikit-learn
tensorflow


# for Gemini
langchain_core
langchain-google-genai # This package contains the LangChain integrations for the Gemini API.